refactor: move note import logic to its own file
