+++ crates/rust-client/src/notes/mod.rs
@@ -1,19 +1,16 @@
 use alloc::{collections::BTreeSet, string::ToString, vec::Vec};
 
-use miden_objects::{
-    accounts::AccountId, assembly::ProgramAst, crypto::rand::FeltRng, transaction::InputNote,
-};
+use miden_objects::{accounts::AccountId, assembly::ProgramAst, crypto::rand::FeltRng};
 use miden_tx::{auth::TransactionAuthenticator, ScriptTarget};
-use tracing::info;
 use winter_maybe_async::{maybe_async, maybe_await};
 
 use crate::{
-    rpc::{NodeRpcClient, NoteDetails},
-    store::{InputNoteRecord, NoteFilter, NoteStatus, OutputNoteRecord, Store, StoreError},
-    sync::get_nullifier_prefix,
+    rpc::NodeRpcClient,
+    store::{InputNoteRecord, NoteFilter, OutputNoteRecord, Store},
     Client, ClientError, IdPrefixFetchError,
 };
 
+mod import;
 mod note_screener;
 
 // RE-EXPORTS
@@ -127,207 +124,6 @@ impl<N: NodeRpcClient, R: FeltRng, S: Store, A: TransactionAuthenticator> Client
             .expect("The vector always has one element for NoteFilter::Unique"))
     }
 
-    // INPUT NOTE CREATION
-    // --------------------------------------------------------------------------------------------
-
-    /// Imports a new input note into the client's store. The information stored depends on the
-    /// type of note file provided.
-    ///
-    /// If the note file is a [NoteFile::NoteId], the note is fecthed from the node and stored in
-    /// the client's store. If the note is private or does not exist, an error is returned. If the
-    /// ID was already stored, the inclusion proof and metadata are updated.
-    /// If the note file is a [NoteFile::NoteDetails], a new note is created with the provided
-    /// details. The note is marked as ignored if it contains no tag or if the tag is not relevant.
-    /// If the note file is a [NoteFile::NoteWithProof], the note is stored with the provided
-    /// inclusion proof and metadata. The MMR data is not fetched from the node.
-    pub async fn import_note(&mut self, note_file: NoteFile) -> Result<NoteId, ClientError> {
-        let note = match note_file {
-            NoteFile::NoteId(id) => {
-                let mut chain_notes = self.rpc_api.get_notes_by_id(&[id]).await?;
-                if chain_notes.is_empty() {
-                    return Err(ClientError::ExistenceVerificationError(id));
-                }
-
-                let note_details =
-                    chain_notes.pop().expect("chain_notes should have at least one element");
-
-                let inclusion_details = note_details.inclusion_details();
-
-                // Add the inclusion proof to the imported note
-                let inclusion_proof = NoteInclusionProof::new(
-                    inclusion_details.block_num,
-                    inclusion_details.note_index,
-                    inclusion_details.merkle_path.clone(),
-                )?;
-
-                let tracked_note = maybe_await!(self.get_input_note(id));
-
-                if let Err(ClientError::StoreError(StoreError::NoteNotFound(_))) = tracked_note {
-                    let node_note = match note_details {
-                        NoteDetails::Public(note, _) => note,
-                        NoteDetails::OffChain(..) => {
-                            return Err(ClientError::NoteImportError(
-                                "Incomplete imported note is private".to_string(),
-                            ))
-                        },
-                    };
-
-                    // If note is not tracked, we create a new one.
-                    let details = node_note.clone().into();
-
-                    let nullifiers = self
-                        .rpc_api
-                        .check_nullifiers_by_prefix(&[get_nullifier_prefix(&node_note.nullifier())])
-                        .await
-                        .map_err(ClientError::RpcError)?;
-
-                    let status = if let Some((_, block_num)) =
-                        nullifiers.iter().find(|(nullifier, _)| nullifier == &node_note.nullifier())
-                    {
-                        NoteStatus::Consumed {
-                            consumer_account_id: None,
-                            block_height: *block_num as u64,
-                        }
-                    } else {
-                        NoteStatus::Committed {
-                            block_height: inclusion_proof.location().block_num() as u64,
-                        }
-                    };
-
-                    InputNoteRecord::new(
-                        node_note.id(),
-                        node_note.recipient().digest(),
-                        node_note.assets().clone(),
-                        status,
-                        Some(*node_note.metadata()),
-                        Some(inclusion_proof),
-                        details,
-                        false,
-                        None,
-                    )
-                } else {
-                    // If note is already tracked, we update the inclusion proof and metadata.
-                    let tracked_note = tracked_note?;
-
-                    // TODO: Join these calls to one method that updates both fields with one query (issue #404)
-                    maybe_await!(self
-                        .store
-                        .update_note_inclusion_proof(tracked_note.id(), inclusion_proof))?;
-                    maybe_await!(self
-                        .store
-                        .update_note_metadata(tracked_note.id(), *note_details.metadata()))?;
-
-                    return Ok(tracked_note.id());
-                }
-            },
-            NoteFile::NoteDetails { details, after_block_num: _, tag: None } => {
-                let record_details = details.clone().into();
-
-                InputNoteRecord::new(
-                    details.id(),
-                    details.recipient().digest(),
-                    details.assets().clone(),
-                    NoteStatus::Expected { created_at: 0 },
-                    None,
-                    None,
-                    record_details,
-                    true,
-                    None,
-                )
-            },
-            NoteFile::NoteDetails { details, after_block_num, tag: Some(tag) } => {
-                let tracked_tags = maybe_await!(self.get_note_tags())?;
-
-                let account_tags = maybe_await!(self.get_account_stubs())?
-                    .into_iter()
-                    .map(|(stub, _)| NoteTag::from_account_id(stub.id(), NoteExecutionMode::Local))
-                    .collect::<Result<Vec<_>, _>>()?;
-
-                let uncommited_note_tags =
-                    maybe_await!(self.get_input_notes(NoteFilter::Expected))?
-                        .into_iter()
-                        .filter_map(|note| note.metadata().map(|metadata| metadata.tag()))
-                        .collect::<Vec<_>>();
-
-                let ignored =
-                    ![tracked_tags, account_tags, uncommited_note_tags].concat().contains(&tag);
-
-                if ignored {
-                    info!("Ignoring note with tag {}", tag);
-                }
-
-                let record_details = details.clone().into();
-
-                if let (NoteStatus::Committed { block_height }, Some(input_note)) =
-                    self.check_expected_note(after_block_num, tag, &details).await?
-                {
-                    let mut current_partial_mmr =
-                        maybe_await!(self.build_current_partial_mmr(true))?;
-                    self.get_and_store_authenticated_block(
-                        block_height.try_into().map_err(|_| {
-                            ClientError::NoteImportError(
-                                "Couldn't convert block height".to_string(),
-                            )
-                        })?,
-                        &mut current_partial_mmr,
-                    )
-                    .await?;
-                    InputNoteRecord::from(input_note)
-                } else {
-                    InputNoteRecord::new(
-                        details.id(),
-                        details.recipient().digest(),
-                        details.assets().clone(),
-                        NoteStatus::Expected { created_at: 0 },
-                        None,
-                        None,
-                        record_details,
-                        ignored,
-                        Some(tag),
-                    )
-                }
-            },
-            NoteFile::NoteWithProof(note, inclusion_proof) => {
-                let details = note.clone().into();
-
-                let nullifiers = self
-                    .rpc_api
-                    .check_nullifiers_by_prefix(&[get_nullifier_prefix(&note.nullifier())])
-                    .await
-                    .map_err(ClientError::RpcError)?;
-
-                let status = if let Some((_, block_num)) =
-                    nullifiers.iter().find(|(nullifier, _)| nullifier == &note.nullifier())
-                {
-                    NoteStatus::Consumed {
-                        consumer_account_id: None,
-                        block_height: *block_num as u64,
-                    }
-                } else {
-                    NoteStatus::Committed {
-                        block_height: inclusion_proof.location().block_num() as u64,
-                    }
-                };
-
-                InputNoteRecord::new(
-                    note.id(),
-                    note.recipient().digest(),
-                    note.assets().clone(),
-                    status,
-                    Some(*note.metadata()),
-                    Some(inclusion_proof),
-                    details,
-                    false,
-                    None,
-                )
-            },
-        };
-        let id = note.id();
-
-        maybe_await!(self.store.insert_input_note(note))?;
-        Ok(id)
-    }
-
     /// Compiles the provided program into a [NoteScript] and checks (to the extent possible) if
     /// the specified note program could be executed against all accounts with the specified
     /// interfaces.
@@ -339,67 +135,6 @@ impl<N: NodeRpcClient, R: FeltRng, S: Store, A: TransactionAuthenticator> Client
         self.tx_executor
             .compile_note_script(note_script_ast, target_account_procs)
             .map_err(ClientError::TransactionExecutorError)
-    }
-
-    // HELPERS
-    // ===============================================================================================
-
-    /// Synchronizes a note with the chain.
-    async fn check_expected_note(
-        &mut self,
-        mut request_block_num: u32,
-        tag: NoteTag,
-        expected_note: &miden_objects::notes::NoteDetails,
-    ) -> Result<(NoteStatus, Option<InputNote>), ClientError> {
-        let current_block_num = maybe_await!(self.get_sync_height())?;
-        loop {
-            if request_block_num > current_block_num {
-                return Ok((NoteStatus::Expected { created_at: 0 }, None));
-            };
-
-            let sync_notes = self.rpc_api().sync_notes(request_block_num, &[tag]).await?;
-
-            if sync_notes.block_header.block_num() == sync_notes.chain_tip {
-                return Ok((NoteStatus::Expected { created_at: 0 }, None));
-            }
-
-            // This means that notes with that note_tag were found.
-            // Therefore, we should check if a note with the same id was found.
-            let committed_note =
-                sync_notes.notes.iter().find(|note| note.note_id() == &expected_note.id());
-
-            if let Some(note) = committed_note {
-                // This means that a note with the same id was found.
-                // Therefore, we should mark the note as committed.
-                let note_block_num = sync_notes.block_header.block_num();
-
-                let note_inclusion_proof = NoteInclusionProof::new(
-                    note_block_num,
-                    note.note_index(),
-                    note.merkle_path().clone(),
-                )?;
-
-                return Ok((
-                    NoteStatus::Committed {
-                        // Block header can't be None since we check that already in the if statement.
-                        block_height: note_block_num as u64,
-                    },
-                    Some(InputNote::authenticated(
-                        Note::new(
-                            expected_note.assets().clone(),
-                            note.metadata(),
-                            expected_note.recipient().clone(),
-                        ),
-                        note_inclusion_proof,
-                    )),
-                ));
-            } else {
-                // This means that a note with the same id was not found.
-                // Therefore, we should request again for sync_notes with the same note_tag
-                // and with the block_num of the last block header (sync_notes.block_header.unwrap()).
-                request_block_num = sync_notes.block_header.block_num();
-            }
-        }
     }
 }
